requirement.txt
fastapi
uvicorn
python-dotenv
keybert
pymupdf
python-docx
sentence-transformers
scikit-learn
numpy
requests
langchain-ollama
